api_app/requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.1
chromadb==0.5.0
google-cloud-core==2.4.1
google-cloud-aiplatform==1.48.0
tqdm==4.66.2
python-dotenv==1.0.0
uvicorn
fastapi
pydantic
